pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["NIK <mkisel28@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
django-filter = "^24.2"
gunicorn = "^22.0.0"
psycopg = {extras = ["binary"], version = "^3.2.1"}
djangorestframework = "^3.15.2"
black = "^24.4.2"
pytz = "^2024.1"
drf-yasg = "1.21.7"
setuptools = "^71.1.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
django-cors-headers = "^4.4.0"
pillow = "^10.4.0"
dj-rest-auth = "^6.0.0"
django-allauth = {extras = ["socialaccount"], version = "^64.2.1"}
requests = "^2.32.3"
celery = "^5.4.0"
redis = "^5.0.8"
flower = "^2.0.1"
sentry-sdk = {extras = ["django"], version = "^2.15.0"}
django-debug-toolbar = "^4.4.6"
drf-nested-routers = "^0.94.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
